+++ src/bioimageio/core/weight_converters/pytorch_to_onnx.py
@@ -1,14 +1,9 @@
 from pathlib import Path
 
-import torch
-
 from bioimageio.spec.model.v0_5 import ModelDescr, OnnxWeightsDescr
 
-from .. import __version__
 from ..backends.pytorch_backend import load_torch_model
-from ..digest_spec import get_member_id, get_test_input_sample
-from ..proc_setup import get_pre_and_postprocessing
-from ._utils_onnx import get_dynamic_axes
+from ._utils_torch_onnx import export_to_onnx
 
 
 def convert(
@@ -16,7 +11,7 @@ def convert(
     output_path: Path,
     *,
     verbose: bool = False,
-    opset_version: int = 15,
+    opset_version: int = 18,
 ) -> OnnxWeightsDescr:
     """
     Convert model weights from the Torchscript state_dict format to the ONNX format.
@@ -29,14 +24,14 @@ def convert(
         verbose:
             If True, will print out detailed information during the ONNX export process. Defaults to False.
         opset_version:
-            The ONNX opset version to use for the export. Defaults to 15.
+            The ONNX opset version to use for the export. Defaults to 18.
 
     Raises:
         ValueError:
             If the provided model does not have weights in the PyTorch state_dict format.
 
     Returns:
-        A descriptor object that contains information about the exported ONNX weights.
+        A description of the exported ONNX weights.
     """
 
     state_dict_weights_descr = model_descr.weights.pytorch_state_dict
@@ -45,35 +40,13 @@ def convert(
             "The provided model does not have weights in the pytorch state dict format"
         )
 
-    sample = get_test_input_sample(model_descr)
-    procs = get_pre_and_postprocessing(
-        model_descr, dataset_for_initial_statistics=[sample]
-    )
-    procs.pre(sample)
-    inputs_numpy = [
-        sample.members[get_member_id(ipt)].data.data for ipt in model_descr.inputs
-    ]
-    inputs_torch = [torch.from_numpy(ipt) for ipt in inputs_numpy]
     model = load_torch_model(state_dict_weights_descr, load_state=True)
-    with torch.no_grad():
-        outputs_original_torch = model(*inputs_torch)
-        if isinstance(outputs_original_torch, torch.Tensor):
-            outputs_original_torch = [outputs_original_torch]
 
-        _ = torch.onnx.export(
-            model,
-            tuple(inputs_torch),
-            str(output_path),
-            input_names=[str(d.id) for d in model_descr.inputs],
-            output_names=[str(d.id) for d in model_descr.outputs],
-            dynamic_axes=get_dynamic_axes(model_descr),
-            verbose=verbose,
-            opset_version=opset_version,
-        )
-
-    return OnnxWeightsDescr(
-        source=output_path.absolute(),
+    return export_to_onnx(
+        model_descr,
+        model,
+        output_path,
+        verbose,
+        opset_version,
         parent="pytorch_state_dict",
-        opset_version=opset_version,
-        comment=f"Converted with bioimageio.core {__version__}.",
     )